Refactoring AuthenticatedPage header into it's own component
